Fix Labels layer dtype to prevent rendering issues
Use np.int32 instead of int (int64) for mask arrays. napari has
known issues with int64 labels not rendering/painting correctly.

+++ src/napari_mm3/_annotate.py
@@ -293,7 +293,7 @@ class Annotate(MM3Container):
         timestamps = [
             int(get_numbers.findall(filename.name)[0]) for filename in filenames
         ]
-        mask_stack = np.zeros(np.shape(img_stack), dtype=np.uint16)
+        mask_stack = np.zeros(np.shape(img_stack), dtype=np.int32)
         for timestamp, filename in zip(timestamps, filenames):
             with tiff.TiffFile(filename) as tif:
                 mask_stack[timestamp, :, :] = tif.asarray()
